feat: use statistics endpoint to get element instances count
